requirements.txt
Flask==2.0.1
tensorflow-cpu==2.19.0
pandas==1.5.3
numpy==1.26.4
matplotlib==3.7.1
seaborn==0.12.2
sqlalchemy==1.4.31
google-api-python-client==2.37.0
google-auth==2.6.0
scikit-learn==1.3.0
plotly==5.6.0
dash==2.3.0
flask-wtf==1.0.0
gunicorn==20.1.0
pytest==7.0.1
flake8==4.0.1
black==22.1.0
python-dotenv==0.19.2
